Fix PHP Warnings when creating and deleting a device (eqLogic)

+++ core/class/JeedomConnectUtils.class.php
@@ -572,11 +572,12 @@ class JeedomConnectUtils {
      * @return void
      */
     public static function delTree($dir) {
-        $files = array_diff(scandir($dir), array('.', '..'));
-        foreach ($files as $file) {
-            (is_dir("$dir/$file")) ? self::delTree("$dir/$file") : unlink("$dir/$file");
+        $files = @scandir($dir);
+        if ($files === false) return true;
+        foreach (array_diff($files, array('.', '..')) as $file) {
+            (is_dir("$dir/$file")) ? self::delTree("$dir/$file") : @unlink("$dir/$file");
         }
-        return rmdir($dir);
+        return @rmdir($dir);
     }
